Reorder existing code. No actual code changes.
