+++ BaghChal/BaghChal.cs
@@ -7,16 +7,10 @@ using System.Threading.Tasks;
 namespace BaghChal
 {
 
-    public enum Pieces : int
-    {
-        Tiger = 0,
-        Goat = 1,
-        Any = 2,
-        Empty = 3
-    }
-
     public class GameBoard : ICloneable
     {
+        #region Properties and fields
+
         /// <summary>
         /// Holds the state of the board.
         /// First index is for tigers, goats, and both.
@@ -40,28 +34,9 @@ namespace BaghChal
 
         public int GoatsCaptured { get; private set; } = 0;
 
-        public object Clone()
-        {
-            return new GameBoard(Board, CurrentUsersTurn, Ply, GoatsLeftToPlace, GoatsCaptured);
-        }
+        #endregion
 
-        public override bool Equals(object obj)
-        {
-            // Check for null values and compare run-time types. For now allow child types.
-            if (obj == null || !(obj is GameBoard))
-                return false;
-
-            var p = (GameBoard)obj;
-            return (Board[0] == p.Board[0]) && (Board[1] == p.Board[1])
-                && (GoatsCaptured == p.GoatsCaptured) && GoatsLeftToPlace == p.GoatsLeftToPlace
-                && (Ply == p.Ply);
-        }
-
-        public override int GetHashCode()
-        {
-            // TODO: Not sure if this is a good hash function. Issue added to make it handle transposed positions.
-            return Board[0].GetHashCode() ^ Board[1].GetHashCode();
-        }
+        #region constructors and factories
 
         /// <summary>
         /// Set up a default game with four tigers in the corner.
@@ -106,6 +81,37 @@ namespace BaghChal
             }
         }
 
+        #endregion
+
+        #region interfaces
+
+        public object Clone()
+        {
+            return new GameBoard(Board, CurrentUsersTurn, Ply, GoatsLeftToPlace, GoatsCaptured);
+        }
+
+        #endregion
+
+        #region overrides
+
+        public override bool Equals(object obj)
+        {
+            // Check for null values and compare run-time types. For now allow child types.
+            if (obj == null || !(obj is GameBoard))
+                return false;
+
+            var p = (GameBoard)obj;
+            return (Board[0] == p.Board[0]) && (Board[1] == p.Board[1])
+                && (GoatsCaptured == p.GoatsCaptured) && GoatsLeftToPlace == p.GoatsLeftToPlace
+                && (Ply == p.Ply);
+        }
+
+        public override int GetHashCode()
+        {
+            // TODO: Not sure if this is a good hash function. Issue added to make it handle transposed positions.
+            return Board[0].GetHashCode() ^ Board[1].GetHashCode();
+        }
+
         public override string ToString()
         {
             char[] symbols = { 'T', 'G', '/', '-', };
@@ -122,6 +128,32 @@ namespace BaghChal
             }
             sb.AppendLine("=====================================");
             return sb.ToString();
+        }
+
+        #endregion
+
+        #region methods
+
+        #region Basic game board functionality
+
+        public bool IsPieceAtIndex(Pieces type, int index)
+        {
+            long shiftMe = 1L;
+            return ((shiftMe << index) & Board[(int)type]) != 0L;
+        }
+
+        public Pieces GetPieceAtIndex(int index)
+        {
+            long shiftMe = 1L;
+            if (((shiftMe << index) & Board[(int)Pieces.Any]) != 0L)
+            {
+                if (((shiftMe << index) & Board[(int)Pieces.Tiger]) != 0L)
+                    return Pieces.Tiger;
+                else
+                    return Pieces.Goat;
+            }
+            else
+                return Pieces.Empty;
         }
 
         /// <summary>
@@ -142,24 +174,110 @@ namespace BaghChal
             Board[(int)Pieces.Any] = Board[(int)Pieces.Any] | shift;
         }
 
-        public bool IsPieceAtIndex(Pieces type, int index)
+        private int[] GetLinkedPositions(int start)
         {
-            long shiftMe = 1L;
-            return ((shiftMe << index) & Board[(int)type]) != 0L;
+            return new int[] {
+                start - 8, start - 7, start - 6,
+                start - 1,            start + 1,
+                start + 6, start + 7, start + 8,
+            };
         }
 
-        public Pieces GetPieceAtIndex(int index)
+        private int[] GetJumpLinkedPositions(int start)
         {
-            long shiftMe = 1L;
-            if (((shiftMe << index) & Board[(int)Pieces.Any]) != 0L) {
-                if (((shiftMe << index) & Board[(int)Pieces.Tiger]) != 0L)
-                    return Pieces.Tiger;
-                else
-                    return Pieces.Goat;
+            // TODO: Create a index for positions instead (left up, up, right up, right, and so on.
+            // move is one step in index direction (-8, -7, -6, +1, ...). Jump is just 2 times that.
+            return new int[] {
+                start - 16, start - 14, start - 12,
+                start -  2,             start +  2,
+                start + 12, start + 14, start + 16,
+            };
+        }
+
+        public static MoveType PositionsAreLinked(int start, int end)
+        {
+            // Skipping check of Out of bounds here. Assume it's done.
+            // Are positions up, down, left, and right are connected.
+            // Assume positive indexes here.
+            var dx = Math.Abs(start - end);
+            if (dx == 1 || dx == 7)
+                return MoveType.Linked;
+            // A jump is simply a move of two steps. Though only tigers can do it and there must be a goat between.
+            // Though this code only checks if it's a possible move from a board perspective.
+            else if (dx == 2 || dx == 14)
+                return MoveType.Jump;
+            // There are also diagonal lines going going from even coordinates (-2, 0, 2, 4) down right
+            // and the same down left (2, 4, 6, 8)
+            // First are multiples of 8. Second case are multiples of 6.
+            // Since this can occure on other lines as well, we have to check that one position is on a line.
+            // All diagonal lines are on even squares.
+            // Fun note. % is not modulus, but reminder, in C#.
+            var even = (start % 2) == 0;
+            if (even && (dx == 6 || dx == 8))
+                return MoveType.Linked;
+            // Diagonal jump.
+            else if (even && (dx == 12 || dx == 16))
+                return MoveType.Jump;
+            else
+                return MoveType.OutOfReach;
+        }
+
+        /// <summary>
+        /// Check that a jump move is valued.
+        /// Can only be done by a tiger.
+        /// Must be done over a goat, into a empty space.
+        /// </summary>
+        private bool IsJumpValid(Pieces piece, int start, int end, out int captureIndex)
+        {
+            captureIndex = -1;
+            if (piece != Pieces.Tiger)
+                return false;
+            // Should be able to omit this and leave it to the caller.
+            if (start == end)
+                return false;
+            if (IsPieceAtIndex(Pieces.Any, end))
+                return false;
+
+            var middleIndex = ((start - end) / 2) + end;
+            if (IsPieceAtIndex(Pieces.Goat, middleIndex) && !IsPieceAtIndex(Pieces.Any, end))
+            {
+                captureIndex = middleIndex;
+                return true;
             }
             else
-                return Pieces.Empty;
+                return false;
         }
+
+        /// <summary>
+        /// Translate a 1 indexed x, y coordinate into a index position on the 
+        /// internal board storage.
+        /// </summary>
+        public static int TranslateToBoardIndex((int x, int y) position)
+        {
+            // TODO: Should rewrite the entire internal code to only work with indexes. Tried a dictionary and it was slower!
+            return position.x + (position.y * 7);
+        }
+
+        /// <summary>
+        /// Translate a 1 indexed x, y coordinate into a index position on the 
+        /// internal board storage.
+        /// </summary>
+        private static (int x, int y) TranslatetFromBoardIndex(int index)
+        {
+            int x = index % 7;
+            int y = index / 7;
+            return (x, y);
+        }
+
+        private bool IsOutOfBounds((int x, int y) position)
+        {
+            return position.x < 1 || position.y < 1 ||
+                position.x > 5 || position.y > 5;
+        }
+
+        #endregion
+
+        #region Game logic
 
         private void Tick()
         {
@@ -172,13 +290,13 @@ namespace BaghChal
             switch (piece)
             {
                 case Pieces.Tiger:
-                    if(checkAfterMove)
+                    if (checkAfterMove)
                         return (GoatsCaptured == 5);
                     return !TigerAbleToMove();
                 case Pieces.Goat:
                     if (checkAfterMove)
                         return !TigerAbleToMove();
-                    return (GoatsCaptured == 5); 
+                    return (GoatsCaptured == 5);
                 default:
                     return false;
             }
@@ -232,7 +350,7 @@ namespace BaghChal
         public Dictionary<(int x, int y), List<(MoveResult result, (int x, int y) positions)>> GetGoatMoves()
         {
             Dictionary<(int x, int y), List<(MoveResult result, (int x, int y) positions)>> res = new Dictionary<(int x, int y), List<(MoveResult result, (int x, int y) posistion)>>();
-            if(GoatsLeftToPlace != 0)
+            if (GoatsLeftToPlace != 0)
             {
                 List<(MoveResult result, (int x, int y) posistion)> returnValue = new List<(MoveResult result, (int x, int y) posistion)>();
                 for (int row = 1; row < 6; row++)
@@ -270,26 +388,6 @@ namespace BaghChal
             return res;
         }
 
-        private int[] GetLinkedPositions(int start)
-        {
-            return new int[] {
-                start - 8, start - 7, start - 6,
-                start - 1,            start + 1,
-                start + 6, start + 7, start + 8,
-            };
-        }
-
-        private int[] GetJumpLinkedPositions(int start)
-        {
-            // TODO: Create a index for positions instead (left up, up, right up, right, and so on.
-            // move is one step in index direction (-8, -7, -6, +1, ...). Jump is just 2 times that.
-            return new int[] {
-                start - 16, start - 14, start - 12,
-                start -  2,             start +  2,
-                start + 12, start + 14, start + 16,
-            };
-        }
-        
         public MoveResult Move(Pieces piece, (int x, int y) start, (int x, int y) end)
         {
             var move = TryMove(piece, start, end);
@@ -365,61 +463,7 @@ namespace BaghChal
                 // All checks should have passed. Now perform normal 1 slot move.
                 return MoveResult.MoveOK;
             }
-            
-        }
 
-        public static MoveType PositionsAreLinked(int start, int end)
-        {
-            // Skipping check of Out of bounds here. Assume it's done.
-            // Are positions up, down, left, and right are connected.
-            // Assume positive indexes here.
-            var dx = Math.Abs(start - end);
-            if (dx == 1 || dx == 7)
-                return MoveType.Linked;
-            // A jump is simply a move of two steps. Though only tigers can do it and there must be a goat between.
-            // Though this code only checks if it's a possible move from a board perspective.
-            else if (dx == 2 || dx == 14)
-                return MoveType.Jump;
-            // There are also diagonal lines going going from even coordinates (-2, 0, 2, 4) down right
-            // and the same down left (2, 4, 6, 8)
-            // First are multiples of 8. Second case are multiples of 6.
-            // Since this can occure on other lines as well, we have to check that one position is on a line.
-            // All diagonal lines are on even squares.
-            // Fun note. % is not modulus, but reminder, in C#.
-            var even = (start % 2) == 0;
-            if (even && (dx == 6 || dx == 8))
-                return MoveType.Linked;
-            // Diagonal jump.
-            else if (even && (dx == 12 || dx == 16))
-                return MoveType.Jump;
-            else
-                return MoveType.OutOfReach;
-        }
-
-        /// <summary>
-        /// Check that a jump move is valued.
-        /// Can only be done by a tiger.
-        /// Must be done over a goat, into a empty space.
-        /// </summary>
-        private bool IsJumpValid(Pieces piece, int start, int end, out int captureIndex)
-        {
-            captureIndex = -1;
-            if (piece != Pieces.Tiger)
-                return false;
-            // Should be able to omit this and leave it to the caller.
-            if (start == end)
-                return false;
-            if (IsPieceAtIndex(Pieces.Any, end))
-                return false;
-
-            var middleIndex = ((start - end) / 2) + end;
-            if (IsPieceAtIndex(Pieces.Goat, middleIndex) && !IsPieceAtIndex(Pieces.Any, end))
-            {
-                captureIndex = middleIndex;
-                return true;
-            }
-            else
-                return false;
         }
 
         private void PerformGameMove(Pieces piece, int startIndex, int endIndex, int captureIndex = -1)
@@ -443,7 +487,7 @@ namespace BaghChal
                 GoatsCaptured++;
             }
             // Place goat.
-            else if(startIndex == -1)
+            else if (startIndex == -1)
             {
                 long shift = (1L << endIndex);
                 Board[(int)Pieces.Goat] = Board[(int)Pieces.Goat] | shift;
@@ -465,58 +509,10 @@ namespace BaghChal
             return;
         }
 
-        /// <summary>
-        /// Translate a 1 indexed x, y coordinate into a index position on the 
-        /// internal board storage.
-        /// </summary>
-        public static int TranslateToBoardIndex((int x, int y) position)
-        {
-            // TODO: Should rewrite the entire internal code to only work with indexes. Tried a dictionary and it was slower!
-            return position.x + (position.y * 7);
-        }
+        #endregion
 
-        /// <summary>
-        /// Translate a 1 indexed x, y coordinate into a index position on the 
-        /// internal board storage.
-        /// </summary>
-        private static (int x, int y) TranslatetFromBoardIndex(int index)
-        {
-            int x = index % 7;
-            int y = index / 7;
-            return (x, y);
-        }
+        #endregion
 
-        private bool IsOutOfBounds((int x, int y) position)
-        {
-            return position.x < 1 || position.y < 1 ||
-                position.x > 5 || position.y > 5;
-        }
-
-    }
-
-    public enum MoveResult : int
-    {
-        MoveOK = 1,
-        OutOfBounds = 2,
-        TryToMoveIncorrectPiece = 3,
-        TargetLocationOccupied = 4,
-        TargetLocationOutOfReach = 5,
-        StartPositionEmpty = 6,
-        NotPlayersTurn = 7,
-        TigerWin = 8,
-        GoatWin = 9,
-        Draw = 10,
-        GoatCaptured = 11,
-        GoatMoveDuringPlacement = 12,
-        GoatPlaced = 13,
-        InvalidJump = 14,
-    }
-
-    public enum MoveType : int
-    {
-        Linked = 1,
-        Jump = 2,
-        OutOfReach = 3,
     }
 
 }
